changes suggested by reuven
