+++ packages/storage/src/item.rs
@@ -1,10 +1,10 @@
+use std::{any::type_name};
+
 use std::marker::PhantomData;
 
-use cosmwasm_std::{Storage, ReadonlyStorage, StdResult};
+use cosmwasm_std::{Storage, ReadonlyStorage, StdResult, StdError};
 use secret_toolkit_serialization::{Serde, Bincode2};
 use serde::{Serialize, de::DeserializeOwned};
-
-use crate::typed_storage::TypedStorage;
 
 /// This storage struct is based on Item from cosmwasm-storage-plus
 pub struct Item<'a, T, Ser = Bincode2>
@@ -27,12 +27,6 @@ impl<'a, T: Serialize + DeserializeOwned, Ser: Serde> Item<'a, T, Ser> {
     }
 }
 
-impl<'a, T: Serialize + DeserializeOwned, Ser: Serde> TypedStorage<T, Ser> for Item<'a, T, Ser> {
-    fn as_slice(&self) -> &[u8] {
-        self.storage_key
-    }
-}
-
 impl<'a, T, Ser> Item<'a, T, Ser>
 where
     T: Serialize + DeserializeOwned,
@@ -43,6 +37,7 @@ where
         self.save_impl(storage, data)
     }
 
+    /// userfacing remove function
     pub fn remove<S: Storage>(&self, storage: &mut S) {
         self.remove_impl(storage);
     }
@@ -72,5 +67,55 @@ where
         let output = action(input)?;
         self.save_impl(storage, &output)?;
         Ok(output)
+    }
+
+    /// Returns StdResult<T> from retrieving the item with the specified key.  Returns a
+    /// StdError::NotFound if there is no item with that key
+    ///
+    /// # Arguments
+    ///
+    /// * `storage` - a reference to the storage this item is in
+    fn load_impl<S: ReadonlyStorage>(&self, storage: &S) -> StdResult<T> {
+        Ser::deserialize(
+            &storage.get(self.as_slice())
+                .ok_or(StdError::not_found(type_name::<T>()))?,
+        )
+    }
+
+    /// Returns StdResult<Option<T>> from retrieving the item with the specified key.  Returns a
+    /// None if there is no item with that key
+    ///
+    /// # Arguments
+    ///
+    /// * `storage` - a reference to the storage this item is in
+    fn may_load_impl<S: ReadonlyStorage>(&self, storage: &S) -> StdResult<Option<T>> {
+        match storage.get(self.as_slice()) {
+            Some(value) => Ser::deserialize(&value).map(Some),
+            None => Ok(None),
+        }
+    }
+
+    /// Returns StdResult<()> resulting from saving an item to storage
+    ///
+    /// # Arguments
+    ///
+    /// * `storage` - a mutable reference to the storage this item should go to
+    /// * `value` - a reference to the item to store
+    fn save_impl<S: Storage>(&self, storage: &mut S, value: &T) -> StdResult<()> {
+        storage.set(self.as_slice(), &Ser::serialize(value)?);
+        Ok(())
+    }
+
+    /// Removes an item from storage
+    ///
+    /// # Arguments
+    ///
+    /// * `storage` - a mutable reference to the storage this item is in
+    fn remove_impl<S: Storage>(&self, storage: &mut S) {
+        storage.remove(self.as_slice());
+    }
+
+    fn as_slice(&self) -> &[u8] {
+        self.storage_key
     }
 }
