+++ packages/storage/src/lib.rs
@@ -1,7 +1,6 @@
+#[doc = include_str!("../Readme.md")]
 pub mod append_store;
 pub mod deque_store;
-pub mod typed_storage;
-pub mod prefixed_typed_storage;
 pub mod item;
 pub mod keymap;
